+++ scripts/ci_before_init_embed.bash
@@ -2,7 +2,7 @@
 
 ROS_DISTRO=$1
 
-export WEBOTS_RELEASE_VERSION=2023b
+export WEBOTS_RELEASE_VERSION=2023b-rev1
 export WEBOTS_OFFSCREEN=1
 export CI=1
 export DEBIAN_FRONTEND=noninteractive